# Using sets to lower time complexity on general pruning