Hi @fbarth,

In this pull request, I:

- Changed all `states` variables from lists to sets, improving the performance of verifying if a state is in states from `O(N)` to `O(1)` (see the example below for a performance comparison).
- Updated the code comments to reflect these changes.

### Performance Comparison

Using the `SumOne(1, '', 100_000)` environment with the `BuscaLargura` algorithm and `pruning='general'`:

**Before:**
```sh
time python tests/SumOne.py
```
Output: `"python tests/SumOne.py  61.15s user 0.07s system 99% cpu 1:01.23 total"`

**After:**
```sh
time python tests/SumOne.py
```
Output: `"python tests/SumOne.py  0.31s user 0.00s system 99% cpu 0.312 total"`

This change significantly reduces the execution time, making the algorithm more efficient.

Please let me know if you would like any adjustments.